discover host IP in New
