+++ modules/host/host.go
@@ -2,10 +2,13 @@ package host
 
 import (
 	"errors"
+	"io"
 	"net"
+	"net/http"
 	"os"
 	"time"
 
+	"github.com/NebulousLabs/Sia/build"
 	"github.com/NebulousLabs/Sia/modules"
 	"github.com/NebulousLabs/Sia/modules/consensus"
 	"github.com/NebulousLabs/Sia/sync"
@@ -36,6 +39,7 @@ type Host struct {
 	wallet      modules.Wallet
 	latestBlock types.BlockID
 
+	myAddr         modules.NetAddress
 	saveDir        string
 	spaceRemaining int64
 	fileCounter    int
@@ -106,13 +110,28 @@ func New(cs *consensus.State, tpool modules.TransactionPool, wallet modules.Wall
 	if err != nil {
 		return
 	}
-	go h.listen()
+	h.myAddr = modules.NetAddress(h.listener.Addr().String())
+
+	// discover external IP (during testing, use the loopback address)
+	var hostname string
+	if build.Release == "testing" {
+		hostname = "::1"
+	} else {
+		hostname, err = getExternalIP()
+		if err != nil {
+			return nil, err
+		}
+	}
+	h.myAddr = modules.NetAddress(net.JoinHostPort(hostname, h.myAddr.Port()))
 
 	err = os.MkdirAll(saveDir, 0700)
 	if err != nil {
 		return
 	}
 	h.load()
+
+	// spawn listener
+	go h.listen()
 
 	h.cs.ConsensusSetSubscribe(h)
 
@@ -135,6 +154,11 @@ func (h *Host) Settings() modules.HostSettings {
 	return h.HostSettings
 }
 
+func (h *Host) Address() modules.NetAddress {
+	// no lock needed; h.myAddr is only set once (in New).
+	return h.myAddr
+}
+
 func (h *Host) Info() modules.HostInfo {
 	lockID := h.mu.RLock()
 	defer h.mu.RUnlock(lockID)
@@ -146,4 +170,21 @@ func (h *Host) Info() modules.HostInfo {
 		NumContracts:     len(h.obligationsByID),
 	}
 	return info
+}
+
+// getExternalIP learns the server's hostname from a centralized service,
+// myexternalip.com.
+func getExternalIP() (string, error) {
+	resp, err := http.Get("http://myexternalip.com/raw")
+	if err != nil {
+		return "", err
+	}
+	defer resp.Body.Close()
+	buf := make([]byte, 64)
+	n, err := resp.Body.Read(buf)
+	if err != nil && err != io.EOF {
+		return "", err
+	}
+	hostname := string(buf[:n-1]) // trim newline
+	return hostname, nil
 }
